libavutil: added hwcontext_amf.
Adds hwcontext_amf, which allows to use shared AMF
context for the encoder, decoder and AMF-based filters,
without copy to the host memory.
It will also allow you to use some optimizations in
the interaction of components (for example, SAV) and make a more
manageable and optimal setup for using GPU devices with AMF
in the case of a fully AMF pipeline.
It will be a significant performance uplift when full AMF pipeline
with filters is used.

This patchset includes only a decoder - encoder pipeline,
but such filters have already been prepared:
1) High quality scaler (SR)
2) Color space convertor & scaler

We also plan to add Compression artefact removal filter in near feature.